refactor amp/bento-embedly-card
