+++ extensions/amp-embedly-card/1.0/amp-embedly-card.js
@@ -3,15 +3,17 @@ import {dict} from '#core/types/object';
 
 import {isExperimentOn} from '#experiments';
 
+import {AmpPreactBaseElement} from '#preact/amp-base-element';
+
 import {userAssert} from '#utils/log';
 
 import {AmpEmbedlyKey, TAG as KEY_TAG} from './amp-embedly-key';
-import {BaseElement} from './base-element';
+import {Component, layoutSizeDefined, props, usesShadowDom} from './element';
 
 /** @const {string} */
 const TAG = 'amp-embedly-card';
 
-class AmpEmbedlyCard extends BaseElement {
+class AmpEmbedlyCard extends AmpPreactBaseElement {
   /** @override */
   init() {
     return dict({
@@ -34,6 +36,18 @@ class AmpEmbedlyCard extends BaseElement {
     return layout == Layout_Enum.RESPONSIVE;
   }
 }
+
+/** @override */
+AmpEmbedlyCard['Component'] = Component;
+
+/** @override */
+AmpEmbedlyCard['props'] = props;
+
+/** @override */
+AmpEmbedlyCard['layoutSizeDefined'] = layoutSizeDefined;
+
+/** @override */
+AmpEmbedlyCard['usesShadowDom'] = usesShadowDom;
 
 AMP.extension(TAG, '1.0', (AMP) => {
   AMP.registerElement(TAG, AmpEmbedlyCard);
